Fixing newlines added after curly brace string index access
Fixes #2018

This fixes my most basic use case where I ran into an issue of extra
new lines, so it only deals with assignment of a string character to
another value (see the test for an example of what I mean).  There may
be other cases where newlines are added for this string index access
syntax, though there are language limitations as to what you can do to
a string offset.