make phar extension optional
the utility itself does not need the phar extension if in use as source
version but it previously crashed when invoked w/o the phar extension b/c
there was an unchecked use while trying to map the phar.

fix is to check for the phar class and only if it exists try with phar
mapping. otherwise just take the source version as it is not a phar anyway.