undefined order of files (tppka patch 4), fixes #3171
non invasive fix to deal w/ an inconsistency which is to adopt test
fixtures for undefined order of file result display on standard output
created by the file fixing iteration within the integration test.

props @keradus comment [1] to not rely on the order of fixing and earlier
[2] to not align iterator's flow just for that.

note: there is a similar approach to this in #3660.

refs:

- [1] https://github.com/FriendsOfPHP/PHP-CS-Fixer/issues/3171#issuecomment-374572917

- [2] https://github.com/FriendsOfPHP/PHP-CS-Fixer/pull/3617#discussion_r175851429

- #3171

- #3608

- #3660