Add PHPDoc support for `fully_qualified_strict_types fixer`
Initial work was done in https://github.com/Jadro007/PHP-CS-Fixer/commit/7acec70a16141832a2eef2b29a5cb7446f460766 but since the branch became outdated, it couldn't be just rebased - it required significant changes to make it work again. It wasn't even possible to commit not working version and to provide changes afterwards, because there were huge conflicts and keeping the old code would end up with a mess. I decided to fix this up during rebase, so it's working as it was working in proposed PR (still needs improvements, but these will be done separately).