Add 'functions_opening_brace' option 'next_line_if_return_typehint'
The `functions_opening_brace` option `next_line_unless_newline_at_signature_end`
fixes muti-line function signatures to look like this:

```php
    public function veryLongFoo(
        VeryLongBar $bar,
        VeryLongBaz $baz
    ): VeryLongFoo {
        // function body
    }
```

This PR adds the option `next_line_if_return_typehint` to fix multi-line
function signatures so that the opening brace goes on the next line if a return
typehint is present:

```php
    public function veryLongFoo(
        VeryLongBar $bar,
        VeryLongBaz $baz
    ): VeryLongFoo
    {
        // function body
    }
```

This has the benefit of being visually similar to a single-line function
signature ...

```php
    public function veryLongFoo(VeryLongBar $bar, VeryLongBaz $baz): VeryLongFoo
    {
        // function body
    }
```

... in that the opening brace remains on the next line instead of "jumping" to
the same line, and keeps a (mostly) blank line at the top of the function body.

When combined with the `return_type_declaration` option `'space_before' => 'one'`
with 4-space indentation, this has the further benefit of aligning the parameter
types with the return type:

```php
    public function veryLongFoo(
        VeryLongBar $bar,
        VeryLongBaz $baz
    ) : VeryLongFoo
    {
        // function body
    }
```

That final combination is what I and several teams I have worked with prefer;
this PR allows PHP-CS-Fixer to apply it.