fix: consider method modifiers for `statement_indentation`
If a method is prefixed with any [method modifier][1], the current logic
for handling the start of a `block_signature` scope will not use the
correct indentation, instead using the calculated indentation of the
`T_FUNCTION` token.

This keeps track of the calculated indentation when we encounter such a
modifier, and moves handling a `T_FUNCTION` token to a separate flow that
uses these stored values (walking backwards from the current token) as the
basis for the indentation value of the scope.

[1]: https://github.com/php/php-langspec/blob/master/spec/19-grammar.md#grammar-method-modifier