Rename column `Inherited` in `In at least one set`.
This is to prevent confusion: reading "Inherited" makes think it relates to the current configuration, or that the fixer is inherited in some ways.

This is technically true from the perspective of PHP CS FIxer, but it isn't from the developer's perspective.

If the fixer is not explicitly configured nor is present in any of the sets imported by the developer, this column is anyway true (green): this causes confusion, as the brain thinks "this is active in some ways", but it actually may not be active as the "green" only indicates that the fixer is in a set.

But the fact it is in a set doesn't automatically mean it is also activated.