Use PHP 8.2 migration rule within PHP 8.3 syntax
Currently we don't have dedicated rules for 8.3, but we can take advantage of `simple_to_complex_string_variable` rule which is part of `@PHP82Migration` ruleset and apply it inside dynamic const fetch syntax.

In the future we should have fixer for enforcing const types, which should be added to `@PHP83Migration` ruleset.