Change approach for deprecated rulesets
Only rulesets that are actually invalid and must be removed should be marked as deprecated. In our case it's `@PER` and `@PER:risky` because the name was invalid and they were re-introduced as `@PER-CS` and `@PER-CS:risky`.

Additionally, the inheritance tree was changed and `@PER` now extends `@PER-CS` instead of `@PER-CS2.0` (deprecated alias ruleset pointing to proper alias ruleset).