Import only valid FQNs
If symbol does not start with `\` it means that it's already imported (as a full import or partial like `use Foo\Bar;` + `Bar\Baz` usage), or it's a global symbol in a global namespace. Either way, it's tests' and static analysis' responsibility to catch invalid symbols, we just treat them as valid ones and handle only FQCNs.