Register hidden `WorkerCommand`
Empty for now, because logic for running fixers must be moved here. Basically it's a PoC of command being hidden - works on PHP 7.4 and 8.3 with latest Composer packages on each version.

Command's arguments/options are more or less what we need in the worker, but it may be changed later.