fix: Add initial in-statements indent support
Previously, statements (which here means anything that isn't a block or
an array) were not fixed but adjusted as a whole, meaning that instead
of forcing the indentation level on each line of a given statement, only
the first line was fixed, and subsequent lines were changed with the
same delta.

For example, in the following snippet, line 3 is indented with one more
space than line 2. Line 2 was fixed by adding two spaces, so line 3 was
adjusted with 2 more spaces as well.

```diff
 if (
-  true
-   || false) {}
+    true
+     || false) {}
```