Refactor `OrderedAttributesFixer`
The goal of this refactor is removing indirectness around namespace analysis. Previously, namespace analysis and imports' analysis was done in void, it produced the result as a side effect which was stored in class' property and used later in the code. There was also logic for clearing these analyses, which was invoked at the end of processing of each attribute. That means the order of operations was important.

After refactoring the analyses are a product of a method, that is used directly.