Move soft-reserved keyword resource to own group in phpdoc_types rule
Users that have defined or are using a class `Resource` in their code now have the option to preserve proper casing.

Fixes https://github.com/PHP-CS-Fixer/PHP-CS-Fixer/issues/8095 as suggested in https://github.com/PHP-CS-Fixer/PHP-CS-Fixer/issues/8095#issuecomment-2639562550.

This change maintains the current behaviour when the default configuration is used.
It is backwards compatible when the configuration is customized.