test: Prove bugs in PhpdocOrderFixer when phpstan- order is specified
In v3.77.0, PhpdocOrderFixer added default support for ordering
phpstan- and psalm- annotations after any of the equivalent non-prefixed
tags that are specified.

However, this produces incorrect diffs (unexpectedly reordering tags of
the same type, and entirely removing multiline returns) if the user has
specified an explicit order for these annotations.

This commit proves the issue, to be followed by a separate fix.