fix: PhpdocOrderFixer when phpstan- / psalm- order is specified
In v3.77.0, PhpdocOrderFixer added default support for ordering
phpstan- and psalm- annotations after any of the equivalent non-prefixed
tags that are specified.

If the user had already specified any of these tags in the order, the
implementation was adding duplicates. E.g. a configuration like
`['param', 'phpstan-param', 'return']` was internally processed as
`['param', 'phpstan-param', 'phpstan-param', 'return']`.

The duplicates were then causing annotations to be moved multiple times,
in some cases resulting in them being deleted from the output (as proved
by the testcase in the previous commit).