Mark `DocumentationCommandTest::testGeneratingDocumentation()` as large test
This test is failing in a lot of jobs, wasting people's time on re-running them multiple times. Let's allow this test to work longer and then maybe improve it somehow.

It takes ~5s locally for me (within PHPStorm), so maybe we just hit the point where we have too many files to process and generate docs for.